Remove ticket numbers from all vsn.mk files
Some application's vsn.mk files contained a list of the ticket
numbers fixed in each version.

Since that information can be obtained from the notes.xml file or
from the merge commits in the git repository (provided that the
branch name includes the ticket number), there is no reason to
manually maintain that information in the vsn.mk files.