Prevent xmerl from over-normalizing character references in attributes
Section 3.3.3 of the XML Recommendation gives the rules for
attribute-value normalization.  One of those rules requires
that character references not be re-normalized after being
replaced with the referenced characters:

    For a character reference, append the referenced
    character to the normalized value.

And, in particular:

    Note that if the unnormalized attribute value contains
    a character reference to a white space character other
    than space (#x20), the normalized value contains the
    referenced character itself (#xD, #xA or #x9).

    Source:  http://www.w3.org/TR/xml/#AVNormalize

In xmerl_scan, however, character references in attributes are
normalized an extra time after replacement.  For example, the
character reference "&#xA" in the following XML document gets
normalized (incorrectly) into a space when parsed:

    2> xmerl_scan:string("<root x='&#xA;'/>").
    {... [{xmlAttribute,x,[],[],[],[],1,[]," ",false}] ...}

This short patch restores the correct behavior:

    2> xmerl_scan:string("<root x='&#xA;'/>").
    {... [{xmlAttribute,x,[],[],[],[],1,[],"\n",false}] ...}

NOTE:  This change does not include tests because I could not
find a test suite for xmerl.